Set scratchFile back to blank and create getTicker.py
